deprecate old Coordinates functionality
