+++ pyfar/classes/coordinates.py
@@ -113,102 +113,35 @@ from scipy.spatial.transform import Rotation as sp_rot
 import re
 from copy import deepcopy
 import warnings
-from pyfar.classes.warnings import PyfarDeprecationWarning
 
 import pyfar as pf
 
 
 class Coordinates():
     """
-    This function will be changed in pyfar 0.8.0 and will just be able to
-    get cartesian coordinates. If you want to initialize in an other
-    domain use :py:func:`from_spherical_colatitude`,
+    Create Coordinates object with or without cartesian coordinate points.
+
+    For initialization with non-cartesian points use
+    :py:func:`from_spherical_colatitude`,
     :py:func:`from_spherical_elevation`, :py:func:`from_spherical_front`,
     :py:func:`from_spherical_side`, or :py:func:`from_cylindrical`
     instead. For conversions from or into degree
     use :py:func:`deg2rad` and :py:func:`rad2deg`.
 
-    Create :py:func:`Coordinates` object with or without coordinate points.
-    The points that enter the Coordinates object are defined by the `name`
-    (`domain`, `convention`, and `unit`. The `unit` will be deprecated in pyfar
-    v0.8.0 in favor of fixed default units, see :ref:`coordinate_systems` and
-    :ref:`coordinates`)
-
-    +--------------------+----------+------------+----------+----------+
-    | domain, convention | points_1 | points_2   | points_3 | unit     |
-    +====================+==========+============+==========+==========+
-    | cart, right        | x        | y          | z        | met      |
-    +--------------------+----------+------------+----------+----------+
-    | sph, top_colat     | azimuth  | colatitude | radius   | rad, deg |
-    +--------------------+----------+------------+----------+----------+
-    | sph, top_elev      | azimuth  | elevation  | radius   | rad, deg |
-    +--------------------+----------+------------+----------+----------+
-    | sph, side          | lateral  | polar      | radius   | rad, deg |
-    +--------------------+----------+------------+----------+----------+
-    | sph, front         | phi      | theta      | radius   | rad, deg |
-    +--------------------+----------+------------+----------+----------+
-    | cyl, top           | azimuth  | z          | radius_z | rad, deg |
-    +--------------------+----------+------------+----------+----------+
-
     Parameters
     ----------
-    points_1 : array like, number
-        Points for the first coordinate.
-        ``'points_1'``, ``'points_2'``, and ``'points_3'`` will be renamed
-        to ``'x'``, ``'y'`` and ``'z'`` in pyfar 0.8.0.
-    points_2 : array like, number
-        Points for the second coordinate.
-        ``'points_1'``, ``'points_2'``, and ``'points_3'`` will be renamed
-        to ``'x'``, ``'y'`` and ``'z'`` in pyfar 0.8.0.
-    points_3 : array like, number
-        Points for the third coordinate.
-        ``'points_1'``, ``'points_2'``, and ``'points_3'`` will be renamed
-        to ``'x'``, ``'y'`` and ``'z'`` in pyfar 0.8.0.
-    domain : string
-        ``'domain'``, ``'unit'`` and ``'convention'`` initialization
-        parameters will be deprecated in pyfar 0.8.0 in favor of
-        ``from_*``.  Different units are no longer supported. The unit is
-        meter for distances and radians for angles.
-        domain of the coordinate system
-
-        ``'cart'``
-            Cartesian
-        ``'sph'``
-            Spherical
-        ``'cyl'``
-            Cylindrical
-
-        The default is ``'cart'``.
-    convention: string
-        ``'domain'``, ``'unit'`` and ``'convention'`` initialization
-        parameters will be deprecated in pyfar 0.8.0 in favor of
-        ``from_*``.  Different units are no longer supported.
-        Default angle unit is radiant.
-
-        Coordinate convention (see above)
-        The default is ``'right'`` if domain is ``'cart'``,
-        ``'top_colat'`` if domain is ``'sph'``, and ``'top'`` if domain is
-        ``'cyl'``.
-    unit: string
-        ``'domain'``, ``'unit'`` and ``'convention'`` initialization
-        parameters will be deprecated in pyfar 0.8.0 in favor of
-        ``from_*``. Different units are no longer supported. Default
-        angle unit is radiant.
-        The ``'deg'`` parameter will be deprecated in pyfar 0.8.0 in favor
-        of the :py:func:`deg2rad` and :py:func:`rad2deg`.
-
-        Unit of the coordinate system. By default the first available unit
-        is used, which is meters (``'met'``) for ``domain = 'cart'`` and
-        radians (``'rad'``) in all other cases (See above).
+    x : array like, number
+        The x-coordinates. Note that the x-coordinates must be broadcastable to
+        `y` and `z`.
+    y : array like, number
+        The y-coordinates. Note that the y-coordinates must be broadcastable to
+        `x` and `z`.
+    z : array like, number
+        The z-coordinates. Note that the z-coordinates must be broadcastable to
+        `x` and `y`.
     weights: array like, number, optional
         Weighting factors for coordinate points. The `shape` of the array
         must match the `shape` of the individual coordinate arrays.
-        The default is ``None``.
-    sh_order : int, optional
-        This property will be deprecated in pyfar 0.8.0 in favor of
-        :py:class:`spharpy.samplings.SamplingSphere`
-
-        Maximum spherical harmonic order of the sampling grid.
         The default is ``None``.
     comment : str, optional
         Comment about the stored coordinate points. The default is
@@ -223,45 +156,20 @@ class Coordinates():
     _system: dict = None
 
     def __init__(
-            self, points_1: np.array = np.asarray([]),
-            points_2: np.array = np.asarray([]),
-            points_3: np.array = np.asarray([]),
-            domain: str = 'cart', convention: str = None, unit: str = None,
+            self, x: np.array = np.asarray([]),
+            y: np.array = np.asarray([]),
+            z: np.array = np.asarray([]),
             weights: np.array = None, sh_order=None,
             comment: str = "") -> None:
 
         # init empty object
         super(Coordinates, self).__init__()
 
-        # test Deprecation warning
-        if domain != 'cart' or convention is not None or unit is not None:
-            warnings.warn((
-                "This function will be changed in pyfar 0.8.0 to "
-                "init(x, y, z)."),
-                    PyfarDeprecationWarning)
-
-        # set the coordinate system
-        system = self._make_system(domain, convention, unit)
-        self._system = system
-
         # set coordinates according to system
-        if domain == 'cart':
-            self._set_points(points_1, points_2, points_3)
-        elif domain == 'sph':
-            self._set_sph(
-                points_1, points_2, points_3,
-                system['convention'], system['unit'])
-        elif domain == 'cyl':
-            self._set_cyl(
-                points_1, points_2, points_3,
-                system['convention'], system['unit'])
-        else:
-            raise ValueError(
-                f"Domain for {domain} is not implemented.")
+        self._set_points(x, y, z)
 
         # save meta data
         self._set_weights(weights)
-        self.sh_order = sh_order
         self._comment = comment
 
     @classmethod
@@ -509,531 +417,6 @@ class Coordinates():
         x, y, z = cyl2cart(azimuth, z, rho)
         return cls(x, y, z, weights=weights, comment=comment)
 
-    def set_cart(self, x, y, z, convention='right', unit='met'):
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of :py:func:`cartesian`, :py:func:`x`, :py:func:`y` or :py:func:`z`.
-        Enter coordinate points in cartesian coordinate systems.
-
-        The points that enter the Coordinates object are defined by the
-        `domain`, `convention`, and `unit`
-
-        +--------------------+----------+------------+----------+----------+
-        | domain, convention | points_1 | points_2   | points_3 | unit     |
-        +====================+==========+============+==========+==========+
-        | cart, right        | x        | y          | z        | met      |
-        +--------------------+----------+------------+----------+----------+
-
-        For more information run
-
-        >>> coords = Coordinates()
-        >>> coords.systems()
-
-        Parameters
-        ----------
-        x, y, z: array like, float
-            Points for the first, second, and third coordinate
-        convention : string, optional
-            Convention in which the coordinate points are stored. The default
-            is ``'right'``.
-        unit : string, optional
-            Unit in which the coordinate points are stored. The default is
-            ``'met'`` for meters.
-        """
-
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of .cart, .x, .y or .z."),
-                PyfarDeprecationWarning)
-
-        # set the coordinate system
-        self._system = self._make_system('cart', convention, unit)
-
-        # save coordinates to self
-        self._set_cart(x, y, z)
-
-    def _set_cart(self, x, y, z, convention='right', unit='met'):
-        if convention != 'right':
-            # Can not be tested. Will only be raised if a coordinate system
-            # is not fully implemented.
-            raise ValueError(
-                (f"Conversion for {convention} is not implemented."))
-
-        if unit != 'met':
-            # Can not be tested. Will only be raised if a coordinate system
-            # is not fully implemented.
-            raise ValueError(
-                f"Unit for {unit} is not implemented")
-
-        # make array
-        x = np.atleast_1d(np.asarray(x, dtype=np.float64))
-        y = np.atleast_1d(np.asarray(y, dtype=np.float64))
-        z = np.atleast_1d(np.asarray(z, dtype=np.float64))
-
-        # squeeze
-        if len(x.shape) == 2 and (x.shape[0] == 1 or x.shape[1] == 1):
-            x = x.flatten()
-        if len(y.shape) == 2 and (y.shape[0] == 1 or y.shape[1] == 1):
-            y = y.flatten()
-        if len(z.shape) == 2 and (z.shape[0] == 1 or z.shape[1] == 1):
-            z = z.flatten()
-
-        # save coordinates to self
-        self._set_points(x, y, z)
-
-    def get_cart(self, convention='right', unit='met', convert=False):  # noqa: ARG002
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of :py:func:`cartesian`
-        Get coordinate points in cartesian coordinate systems.
-
-        The points that are returned are defined by the `domain`, `convention`,
-        and `unit`:
-
-        +--------------------+----------+------------+----------+----------+
-        | domain, convention | p[...,1] | p[...,1]   | p[...,1] | units    |
-        +====================+==========+============+==========+==========+
-        | cart, right        | x        | y          | z        | met      |
-        +--------------------+----------+------------+----------+----------+
-
-        For more information run
-
-        >>> coords = Coordinates()
-        >>> coords.systems()
-
-        Parameters
-        ----------
-        convention : string, optional
-            Convention in which the coordinate points are stored. The default
-            is ``'right'``.
-        unit : string, optional
-            Unit in which the coordinate points are stored. The default is
-           ``'met'``.
-        convert : boolean, optional
-            If True, the internal representation of the samplings points will
-            be converted to the queried coordinate system. The default is
-            ``False``, i.e., the internal presentation remains as it is.
-
-        Returns
-        -------
-        points : numpy array
-            Coordinate points. ``points[...,0]`` holds the points for the first
-            coordinate, ``points[...,1]`` the points for the second, and
-            ``points[...,2]`` the points for the third coordinate.
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of .cartesian"),
-                PyfarDeprecationWarning)
-
-        self._system = self._make_system('cart', convention, unit)
-        return self.cartesian
-
-    def set_sph(
-            self, angles_1, angles_2, radius,
-            convention='top_colat', unit='rad'):
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the ``spherical_*`` properties. For conversions from or into degree
-        use :py:func:`deg2rad` and :py:func:`rad2deg`.
-        Enter coordinate points in spherical coordinate systems.
-
-        The points that enter the Coordinates object are defined by the
-        `domain`, `convention`, and `unit`
-
-        +--------------------+----------+------------+----------+----------+
-        | domain, convention | points_1 | points_2   | points_3 | unit     |
-        +====================+==========+============+==========+==========+
-        | sph, top_colat     | azimuth  | colatitude | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-        | sph, top_elev      | azimuth  | elevation  | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-        | sph, side          | lateral  | polar      | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-        | sph, front         | phi      | theta      | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-
-        For more information run
-
-        >>> coords = Coordinates()
-        >>> coords.systems()
-
-        Parameters
-        ----------
-        points_i: array like, number
-            Points for the first, second, and third coordinate
-        convention : string, optional
-            Convention in which the coordinate points are stored. The default
-            is ``'top_colat'``.
-        unit : string, optional
-            Unit in which the coordinate points are stored. The default is
-            ``'rad'``.
-            The ``'deg'`` parameter will be deprecated in pyfar 0.8.0 in favor
-            of the :py:func:`deg2rad` and :py:func:`rad2deg`.
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of the spherical_... properties"),
-                PyfarDeprecationWarning)
-
-        # make array
-        angles_1 = np.atleast_1d(np.asarray(angles_1, dtype=np.float64))
-        angles_2 = np.atleast_1d(np.asarray(angles_2, dtype=np.float64))
-        radius = np.atleast_1d(np.asarray(radius, dtype=np.float64))
-
-        self._set_sph(angles_1, angles_2, radius, convention, unit)
-
-    def _set_sph(
-            self, angles_1, angles_2, radius,
-            convention='top_colat', unit='rad'):
-
-        # Convert to array
-        angles_1 = np.asarray(angles_1)
-        angles_2 = np.asarray(angles_2)
-        radius = np.asarray(radius)
-
-        # squeeze
-        if len(angles_1.shape) == 2 and \
-                (angles_1.shape[0] == 1 or angles_1.shape[1] == 1):
-            angles_1 = angles_1.flatten()
-        if len(angles_2.shape) == 2 and \
-                (angles_2.shape[0] == 1 or angles_2.shape[1] == 1):
-            angles_2 = angles_2.flatten()
-        if len(radius.shape) == 2 and \
-                (radius.shape[0] == 1 or radius.shape[1] == 1):
-            radius = radius.flatten()
-
-        # convert to radians
-        if unit == 'deg':
-            warnings.warn((
-                "'deg' parameter will be deprecated in pyfar 0.8.0 in favor "
-                "of the pyfar.deg2rad and pyfar.rad2deg"),
-                    PyfarDeprecationWarning)
-            angles_1 = angles_1 / 180 * np.pi
-            angles_2 = angles_2 / 180 * np.pi
-
-        # convert to cartesian ...
-        # ... from spherical coordinate systems
-        if convention == 'top_colat':
-            x, y, z = sph2cart(angles_1, angles_2, radius)
-
-        elif convention == 'top_elev':
-            x, y, z = sph2cart(angles_1, np.pi / 2 - angles_2, radius)
-
-        elif convention == 'side':
-            x, z, y = sph2cart(angles_2, np.pi / 2 - angles_1, radius)
-
-        elif convention == 'front':
-            y, z, x = sph2cart(angles_1, angles_2, radius)
-
-        else:
-            # Can not be tested. Will only be raised if a coordinate system
-            # is not fully implemented.
-            raise ValueError(
-                (f"Conversion for {convention} is not implemented."))
-
-        # set the coordinate system
-        self._system = self._make_system('sph', convention, unit)
-
-        # save coordinates to self
-        self._set_points(x, y, z)
-
-    def get_sph(
-            self, convention='top_colat', unit='rad',
-            convert=False):  # noqa: ARG002
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the `spherical_...` properties. For conversions from or into degree
-        use :py:func:`deg2rad` and :py:func:`rad2deg`.
-        Get coordinate points in spherical coordinate systems.
-
-        The points that are returned are defined by the `domain`,
-        `convention`, and `unit`:
-
-        +--------------------+----------+------------+----------+----------+
-        | domain, convention | p[...,1] | p[...,1]   | p[...,1] | units    |
-        +====================+==========+============+==========+==========+
-        | sph, top_colat     | azimuth  | colatitude | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-        | sph, top_elev      | azimuth  | elevation  | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-        | sph, side          | lateral  | polar      | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-        | sph, front         | phi      | theta      | radius   | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-
-        For more information run
-
-        >>> coords = Coordinates()
-        >>> coords.systems()
-
-        Parameters
-        ----------
-        convention : string, optional
-            Convention in which the coordinate points are stored. The default
-            is ``'top_colat'``.
-        unit : string, optional
-            Unit in which the coordinate points are stored. The default is
-            ``'rad'``.
-            The ``'deg'`` parameter will be deprecated in pyfar 0.8.0 in favor
-            of the :py:func:`deg2rad` and :py:func:`rad2deg`.
-        convert : boolean, optional
-            If True, the internal representation of the samplings points will
-            be converted to the queried coordinate system. The default is
-            ``False``, i.e., the internal presentation remains as it is.
-
-        Returns
-        -------
-        points : numpy array
-            Coordinate points. ``points[...,0]`` holds the points for the first
-            coordinate, ``points[...,1]`` the points for the second, and
-            ``points[...,2]`` the points for the third coordinate.
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of the `spherical_*` properties."),
-                PyfarDeprecationWarning)
-
-        if convention == 'top_colat':
-            points = self.spherical_colatitude
-            self._system = self._make_system('sph', 'top_colat', 'rad')
-        elif convention == 'top_elev':
-            points = self.spherical_elevation
-            self._system = self._make_system('sph', 'top_elev', 'rad')
-        elif convention == 'front':
-            points = self.spherical_front
-            self._system = self._make_system('sph', 'front', 'rad')
-        elif convention == 'side':
-            points = self.spherical_side
-            self._system = self._make_system('sph', 'side', 'rad')
-        else:
-            raise ValueError(
-                f"Conversion for {convention} is not implemented.")
-
-        conversion_factor = 1 if unit == 'rad' else 180 / np.pi
-        points[..., 0] = points[..., 0] * conversion_factor
-        points[..., 1] = points[..., 1] * conversion_factor
-        return points
-
-    def _get_sph(
-            self, convention='top_colat', unit='rad',
-            convert=False):  # noqa: ARG002
-        # check if object is empty
-        self._check_empty()
-
-        x = self._x
-        y = self._y
-        z = self._z
-
-        # convert to spherical...
-        # ... top polar systems
-        if convention[0:3] == 'top':
-            angles_1, angles_2, radius = cart2sph(x, y, z)
-            if convention == 'top_elev':
-                angles_2 = np.pi / 2 - angles_2
-
-        # ... side polar system
-        # (idea for simple conversions from Robert Baumgartner and SOFA_API)
-        elif convention == 'side':
-            angles_2, angles_1, radius = cart2sph(x, z, -y)
-            # range angles
-            angles_1 = angles_1 - np.pi / 2
-            angles_2 = np.mod(angles_2 + np.pi / 2, 2 * np.pi) - np.pi / 2
-        # ... front polar system
-        elif convention == 'front':
-            angles_1, angles_2, radius = cart2sph(y, z, x)
-
-        else:
-            raise ValueError(
-                f"Conversion for {convention} is not implemented.")
-
-        # convert to degrees
-        if unit == 'deg':
-            warnings.warn((
-                "'deg' parameter will be deprecated in pyfar 0.8.0 in favor "
-                "of the pyfar.deg2rad and pyfar.rad2deg"),
-                    PyfarDeprecationWarning)
-            angles_1 = angles_1 / np.pi * 180
-            angles_2 = angles_2 / np.pi * 180
-        elif not unit == 'rad':
-            raise ValueError(
-                f"{unit} is not implemented.")
-
-        # return the points
-        return angles_1, angles_2, radius
-
-    def set_cyl(self, azimuth, z, radius_z, convention='top', unit='rad'):
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the :py:func:`cylindrical` property. For conversions from or
-        into degree use :py:func:`deg2rad` and :py:func:`rad2deg`.
-        Enter coordinate points in cylindrical coordinate systems.
-
-        The points that enter the Coordinates object are defined by the
-        `domain`, `convention`, and `unit`
-
-        +--------------------+----------+------------+----------+----------+
-        | domain, convention | points_1 | points_2   | points_3 | unit     |
-        +====================+==========+============+==========+==========+
-        | cyl, top           | azimuth  | z          | radius_z | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-
-        For more information run
-
-        >>> coords = Coordinates()
-        >>> coords.systems()
-
-        Parameters
-        ----------
-        points_i: array like, number
-            Points for the first, second, and third coordinate
-        convention : string, optional
-            Convention in which the coordinate points are stored. The default
-            is ``'top'``.
-        unit : string, optional
-            Unit in which the coordinate points are stored. The default is
-            ``'rad'``.
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of the cylindrical property."),
-                PyfarDeprecationWarning)
-        self._set_cyl(azimuth, z, radius_z, convention, unit)
-
-    def _set_cyl(self, azimuth, z, rho, convention='top', unit='rad'):
-
-        # Convert to array
-        azimuth = np.asarray(azimuth)
-        z = np.asarray(z)
-        rho = np.asarray(rho)
-
-        # squeeze
-        if len(azimuth.shape) == 2 and \
-                (azimuth.shape[0] == 1 or azimuth.shape[1] == 1):
-            azimuth = azimuth.flatten()
-        if len(z.shape) == 2 and \
-                (z.shape[0] == 1 or z.shape[1] == 1):
-            z = z.flatten()
-        if len(rho.shape) == 2 and \
-                (rho.shape[0] == 1 or rho.shape[1] == 1):
-            rho = rho.flatten()
-
-        # convert to radians
-        if unit == 'deg':
-            warnings.warn((
-                "'deg' parameter will be deprecated in pyfar 0.8.0 in favor "
-                "of the pyfar.deg2rad and pyfar.rad2deg"),
-                    PyfarDeprecationWarning)
-            azimuth = azimuth / 180 * np.pi
-        elif not unit == 'rad':
-            raise ValueError(
-                f"{unit} is not implemented.")
-
-        # ... from cylindrical coordinate systems
-        if convention == 'top':
-            x, y, z = cyl2cart(azimuth, z, rho)
-        else:
-            # Can not be tested. Will only be raised if a coordinate system
-            # is not fully implemented.
-            raise ValueError(
-                (f"Conversion for {convention} is not implemented."))
-
-        # set the coordinate system
-        self._system = self._make_system('cyl', convention, unit)
-
-        # save coordinates to self
-        self._set_points(x, y, z)
-
-    def get_cyl(
-            self, convention='top', unit='rad', convert=False):  # noqa: ARG002
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the `cylindrical` property. For conversions from or into degree
-        use :py:func:`deg2rad` and :py:func:`rad2deg`.
-        Get coordinate points in cylindrical coordinate system.
-
-        The points that are returned are defined by the `domain`, `convention`,
-        and `unit`:
-
-        +--------------------+----------+------------+----------+----------+
-        | domain, convention | p[...,1] | p[...,1]   | p[...,1] | units    |
-        +====================+==========+============+==========+==========+
-        | cyl, top           | azimuth  | z          | radius_z | rad, deg |
-        +--------------------+----------+------------+----------+----------+
-
-        For more information run
-
-        >>> coords = Coordinates()
-        >>> coords.systems()
-
-        Parameters
-        ----------
-        convention : string, optional
-            Convention in which the coordinate points are stored. The default
-            is ``'right'``.
-        unit : string, optional
-            Unit in which the coordinate points are stored. The default is
-            ``'rad'``.
-            The ``'deg'`` parameter will be deprecated in pyfar 0.8.0 in favor
-            of the :py:func:`deg2rad` and :py:func:`rad2deg`.
-
-        convert : boolean, optional
-            If True, the internal representation of the samplings points will
-            be converted to the queried coordinate system. The default is
-            False, i.e., the internal presentation remains as it is.
-
-        Returns
-        -------
-        points : numpy array
-            Coordinate points. ``points[...,0]`` holds the points for the first
-            coordinate, ``points[...,1]`` the points for the second, and
-            ``points[...,2]`` the points for the third coordinate.
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of the cylindrical property."),
-                PyfarDeprecationWarning)
-
-        if not convention == 'top':
-            raise ValueError(
-                f"Conversion for {convention} is not implemented.")
-        points = self.cylindrical
-
-
-        conversion_factor = 1 if unit == 'rad' else 180 / np.pi
-        points[..., 0] = points[..., 0] * conversion_factor
-        return points
-
-    def _get_cyl(self, convention='top', unit='rad'):
-        """internal function to convert cart to cyl coordinates"""
-
-        # check if object is empty
-        self._check_empty()
-
-        # convert to cylindrical ...
-        # ... top systems
-        if convention == 'top':
-            azimuth, z, rho = cart2cyl(self.x, self.y, self.z)
-        else:
-            # Can not be tested. Will only be raised if a coordinate system
-            # is not fully implemented.
-            raise ValueError(
-                f"Conversion for {convention} is not implemented.")
-
-        # convert to degrees
-        if unit == 'deg':
-            warnings.warn((
-                "'deg' parameter will be deprecated in pyfar 0.8.0 in favor "
-                "of the pyfar.deg2rad and pyfar.rad2deg"),
-                    PyfarDeprecationWarning)
-            azimuth = azimuth / np.pi * 180
-        elif unit != 'rad':
-            raise ValueError(
-                f"unit for {unit} is not implemented.")
-
-        # return points and convert internal state if desired
-        return azimuth, z, rho
-
     @property
     def weights(self):
         """Get sampling weights."""
@@ -1043,32 +426,6 @@ class Coordinates():
     def weights(self, value):
         """Set sampling weights."""
         self._set_weights(value)
-
-    @property
-    def sh_order(self):
-        """This function will be deprecated in pyfar 0.8.0 in favor
-            of :py:class:`spharpy.samplings.SamplingSphere`.
-            Get the maximum spherical harmonic order."""
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of spharpy.samplings.SamplingSphere."),
-                PyfarDeprecationWarning)
-
-        return self._sh_order
-
-    @sh_order.setter
-    def sh_order(self, value):
-        """This function will be deprecated in pyfar 0.8.0 in favor
-            of :py:class:`spharpy.samplings.SamplingSphere`.
-            Set the maximum spherical harmonic order."""
-
-        if value is not None:
-            warnings.warn((
-                "This function will be deprecated in pyfar 0.8.0 in favor "
-                "of spharpy.samplings.SamplingSphere."),
-                    PyfarDeprecationWarning)
-
-        self._sh_order = int(value) if value is not None else None
 
     @property
     def comment(self):
@@ -1404,98 +761,6 @@ class Coordinates():
         spherical_side = self.spherical_side
         spherical_side[..., 1] = polar
         self.spherical_side = spherical_side
-
-    def systems(self, show='all', brief=False):
-        """
-        This function will be deprecated in pyfar 0.8.0, check the
-        documentation instead.
-        Print coordinate systems and their description on the console.
-
-        .. note::
-           All coordinate systems are described with respect to the right
-           handed cartesian system (``domain='cart'``, ``convention='right'``).
-           Distances are always specified in meters, while angles can be
-           radians or degrees (``unit='rad'`` or ``unit='deg'``).
-
-
-        Parameters
-        ----------
-        show: string, optional
-            ``'current'`` to list the current coordinate system or ``'all'``
-            to list all coordinate systems. The default is ``'all'``.
-        brief : boolean, optional
-            Will only list the domains, conventions and units if True. The
-            default is ``False``.
-
-        Returns
-        -------
-        Prints to console.
-        """
-
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0."),
-                PyfarDeprecationWarning)
-
-        if show == 'current':
-            domain = self._system['domain']
-            convention = self._system['convention']
-            unit = self._system['unit']
-        elif show == 'all':
-            domain = convention = unit = 'all'
-        else:
-            raise ValueError("show must be 'current' or 'all'.")
-
-        # get coordinate systems
-        systems = self._systems()
-
-        # print information
-        domains = list(systems) if domain == 'all' else [domain]
-
-        if brief:
-            print('domain, convention, unit')
-            print('- - - - - - - - - - - - -')
-            for dd in domains:
-                conventions = list(systems[dd]) if convention == 'all' \
-                    else [convention]
-                for cc in conventions:
-                    # current coordinates
-                    coords = systems[dd][cc]['coordinates']
-                    # current units
-                    if unit != 'all':
-                        units = [units for units in systems[dd][cc]['units']
-                                 if unit == units[0][0:3]]
-                    else:
-                        units = systems[dd][cc]['units']
-                    # key for unit
-                    unit_key = [u[0][0:3] for u in units]
-                    print(f"{dd}, {cc}, [{', '.join(unit_key)}]")
-        else:
-            for dd in domains:
-                conventions = \
-                    list(systems[dd]) if convention == 'all' else [convention]
-                for cc in conventions:
-                    # current coordinates
-                    coords = systems[dd][cc]['coordinates']
-                    # current units
-                    if unit != 'all':
-                        units = [units for units in systems[dd][cc]['units']
-                                 if unit == units[0][0:3]]
-                    else:
-                        units = systems[dd][cc]['units']
-                    # key for unit
-                    unit_key = [u[0][0:3] for u in units]
-                    print("- - - - - - - - - - - - - - - - - "
-                          "- - - - - - - - - - - - - - - - -")
-                    print(f"domain: {dd}, convention: {cc}, unit: "
-                          f"[{', '.join(unit_key)}]\n")
-                    print(systems[dd][cc]['description_short'] + '\n')
-                    print("Coordinates:")
-                    for nn, coord in enumerate(coords):
-                        cur_units = [u[nn] for u in units]
-                        print(
-                            f"points_{nn + 1}: {coord} ",
-                            f"[{', '.join(cur_units)}]")
-                    print('\n' + systems[dd][cc]['description'] + '\n\n')
 
     def show(self, mask=None, **kwargs):
         """
@@ -1872,349 +1137,6 @@ class Coordinates():
 
         return index
 
-    def find_nearest_k(self, points_1, points_2, points_3, k=1,
-                       domain='cart', convention='right', unit='met',
-                       show=False):
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the ``find_nearest`` method.
-
-        Find the k nearest coordinates points.
-
-        Parameters
-        ----------
-        points_i : array like, number
-            First, second and third coordinate of the points to which the
-            nearest neighbors are searched.
-        k : int, optional
-            Number of points to return. k must be > 0. The default is ``1``.
-        domain : string, optional
-            Domain of the points. The default is ``'cart'``.
-        convention: string, optional
-            Convention of points. The default is ``'right'``.
-        unit : string, optional
-            Unit of the points. The default is ``'met'`` for meters.
-        show : bool, optional
-            Show a plot of the coordinate points. The default is ``False``.
-
-        Returns
-        -------
-        index : numpy array of ints
-            The locations of the neighbors in the getter methods (e.g.,
-            ``self.cartesian``). Dimension according to `distance` (see below).
-            Missing neighbors are indicated with ``csize``. Also see Notes
-            below.
-        mask : boolean numpy array
-            Mask that contains ``True`` at the positions of the selected points
-            and ``False`` otherwise. Mask is of shape ``cshape``.
-
-        Notes
-        -----
-        :py:class:`scipy.spatial.cKDTree` is used for the search, which
-        requires an (N, 3) array. The coordinate points in self are thus
-        reshaped to (`csize`, 3) before they are passed to ``cKDTree``.
-        The index that  is returned refers to the reshaped coordinate points.
-        To access the points for example use
-
-        >>> points_reshaped = self.cartesian.reshape((self.csize, 3))
-        >>> points_reshaped[index]
-
-        Examples
-        --------
-
-        Find the nearest point in a line
-
-        .. plot::
-
-            >>> import pyfar as pf
-            >>> coords = pf.Coordinates(np.arange(-5, 5), 0, 0)
-            >>> result = coords.find_nearest_k(0, 0, 0, show=True)
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of find_nearest method."),
-                PyfarDeprecationWarning)
-
-        # check the input
-        assert isinstance(k, int) and k > 0 and k <= self.csize, \
-            "k must be an integer > 0 and <= self.csize."
-
-        # get the points
-        _, index, mask = self._find_nearest(
-            points_1, points_2, points_3,
-            domain, convention, unit, show, k, 'k')
-
-        return index, mask
-
-    def find_nearest_cart(self, points_1, points_2, points_3, distance,
-                          domain='cart', convention='right', unit='met',
-                          show=False, atol=1e-15):
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the ``find_within`` method.
-        Find coordinates within a certain distance in meters to query points.
-
-        Parameters
-        ----------
-        points_i : array like, number
-            First, second and third coordinate of the points to which the
-            nearest neighbors are searched.
-        distance : number
-            Euclidean distance in meters in which the nearest points are
-            searched. Must be >= 0.
-        domain : string, optional
-            Domain of the points. The default is ``'cart'``.
-        convention: string, optional
-            Convention of points. The default is ``'right'``.
-        unit : string, optional
-            Unit of the points. The default is ``'met'`` for meters.
-        show : bool, optional
-            Show a plot of the coordinate points. The default is ``False``.
-        atol : float, optional
-            A tolerance that is added to `distance`. The default is`` 1e-15``.
-
-        Returns
-        -------
-        index : numpy array of ints
-            The locations of the neighbors in the getter methods (e.g.,
-            ``cartesian``). Dimension as in :py:func:`~find_nearest_k`.
-            Missing neighbors are indicated with ``csize``. Also see Notes
-            below.
-        mask : boolean numpy array
-            Mask that contains ``True`` at the positions of the selected points
-            and ``False`` otherwise. Mask is of shape ``cshape``.
-
-        Notes
-        -----
-        :py:class:`scipy.spatial.cKDTree` is used for the search, which
-        requires an
-        (N, 3) array. The coordinate points in self are thus reshaped to
-        (`csize`, 3) before they are passed to ``cKDTree``. The index that
-        is returned refers to the reshaped coordinate points. To access the
-        points for example use
-
-        >>> points_reshaped = self.cartesian.reshape((self.csize, 3))
-        >>> points_reshaped[index]
-
-        Examples
-        --------
-
-        Find frontal points within a distance of 0.5 meters
-
-        .. plot::
-
-            >>> import pyfar as pf
-            >>> coords = pf.Coordinates(np.arange(-5, 5), 0, 0)
-            >>> result = coords.find_nearest_cart(2, 0, 0, 0.5, show=True)
-
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of find_within method."),
-                PyfarDeprecationWarning)
-
-        # check the input
-        assert distance >= 0, "distance must be >= 0"
-
-        # get the points
-        distance, index, mask = self._find_nearest(
-            points_1, points_2, points_3,
-            domain, convention, unit, show, distance, 'cart', atol)
-
-        return index, mask
-
-    def find_nearest_sph(self, points_1, points_2, points_3, distance,
-                         domain='sph', convention='top_colat', unit='rad',
-                         show=False, atol=1e-15):
-        """
-        This function will be deprecated in pyfar 0.8.0 in favor
-        of the ``find_within`` method.
-        Find coordinates within certain angular distance to the query points.
-
-        Parameters
-        ----------
-        points_i : array like, number
-            First, second and third coordinate of the points to which the
-            nearest neighbors are searched.
-        distance : number
-            Great circle distance in degrees in which the nearest points are
-            searched. Must be >= 0 and <= 180.
-        domain : string, optional
-            Domain of the input points. The default is ``'sph'``.
-        convention: string, optional
-            Convention of the input points. The default is ``'top_colat'``.
-        unit: string, optional
-            Unit of the input points. The default is ``'rad'``.
-        show : bool, optional
-            Show a plot of the coordinate points. The default is ``False``.
-        atol : float, optional
-            A tolerance that is added to `distance`. The default is ``1e-15``.
-
-        Returns
-        -------
-        index : numpy array of ints
-            The locations of the neighbors in the getter methods (e.g.,
-            ``cartesian``). Dimension as in :py:func:`~find_nearest_k`.
-            Missing neighbors are indicated with ``csize``. Also see Notes
-            below.
-        mask : boolean numpy array
-            Mask that contains ``True`` at the positions of the selected points
-            and ``False`` otherwise. Mask is of shape ``cshape``.
-
-        Notes
-        -----
-        :py:class:`scipy.spatial.cKDTree` is used for the search, which
-        requires an
-        (N, 3) array. The coordinate points in self are thus reshaped to
-        (`csize`, 3) before they are passed to ``cKDTree``. The index that
-        is returned refers to the reshaped coordinate points. To access the
-        points for example use
-
-        ``points_reshaped = points.cartesian.reshape((points.csize, 3))``
-        ``points_reshaped[index]``
-
-        Examples
-        --------
-
-        Find top points within a distance of 45 degrees
-
-        .. plot::
-
-            >>> import pyfar as pf
-            >>> import numpy as np
-            >>> coords = pf.Coordinates.from_spherical_elevation(
-            >>>     0, np.arange(-90, 91, 10)*np.pi/180, 1)
-            >>> result = coords.find_nearest_sph(0, np.pi/2, 1, 45, show=True)
-
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0 in favor "
-            "of find_within method."),
-                PyfarDeprecationWarning)
-
-        # check the input
-        assert distance >= 0 and distance <= 180, \
-            "distance must be >= 0 and <= 180."
-
-        # get radius and check for equality
-        radius = self.radius
-        delta_radius = np.max(radius) - np.min(radius)
-        if delta_radius > 1e-15:
-            raise ValueError(
-                "find_nearest_sph only works if all points have the same \
-                radius. Differences are larger than 1e-15")
-
-        # get the points
-        distance, index, mask = self._find_nearest(
-            points_1, points_2, points_3,
-            domain, convention, unit, show, distance, 'sph', atol,
-            np.max(radius))
-
-        return index, mask
-
-    def find_slice(self, coordinate: str, unit: str, value, tol=0,
-                   show=False, atol=1e-15):
-        """
-        This function will be deprecated in pyfar 0.8.0. Use properties and
-        slicing instead, e.g. ``coords = coords[coords.azimuth>=np.pi]``.
-
-        Find a slice of the coordinates points.
-
-        Parameters
-        ----------
-        coordinate : str
-            Coordinate for slicing.
-        unit : str
-            Unit in which the value is passed
-        value : number
-            Value of the coordinate around which the points are sliced.
-        tol : number, optional
-            Tolerance for slicing. Points are sliced within the range
-            ``[value-tol, value+tol]``. The default is ``0``.
-        show : bool, optional
-            Show a plot of the coordinate points. The default is ``False``.
-        atol : number, optional
-            A tolerance that is added to `tol`. The default is ``1e-15``.
-
-        Returns
-        -------
-        index : tuple of numpy arrays
-            The indices of the selected points as a tuple of arrays. The length
-            of the tuple matches :py:func:`~cdim`. The length of each array
-            matches the number of selected points.
-        mask : boolean numpy array
-            Mask that contains True at the positions of the selected points and
-            False otherwise. Mask is of shape self.cshape.
-
-        Notes
-        -----
-        `value` must be inside the range of the coordinate (see ``.systems``).
-        However, `value` +/- `tol` may exceed the range.
-
-        Examples
-        --------
-
-        Find horizontal slice of spherical coordinate system within a ring of
-        +/- 10 degrees
-
-        .. plot::
-
-            >>> import pyfar as pf
-            >>> import numpy as np
-            >>> coords = pf.Coordinates.from_spherical_elevation(
-            >>>     np.arange(-30, 30, 5)*np.pi/180, 0, 1)
-            >>> result = coords.find_slice('azimuth', 'deg', 0, 5, show=True)
-
-        """
-        warnings.warn((
-            "This function will be deprecated in pyfar 0.8.0. Use properties"
-            " and slicing instead."),
-                PyfarDeprecationWarning)
-
-        # check if the coordinate and unit exist
-        domain, convention, index = self._exist_coordinate(coordinate, unit)
-
-        # get type and range of coordinate
-        c_info = self._systems()[domain][convention][coordinate]
-
-        # convert input to radians
-        value = value / 180 * np.pi if unit == 'deg' else value
-        tol = tol / 180 * np.pi if unit == 'deg' else tol
-
-        # check if  value is within the range of coordinate
-        if c_info[0] in ["bound", "cyclic"]:
-            assert c_info[1][0] <= value <= c_info[1][1], \
-                f"'value' is {value} but must be in the range {c_info[1]}."
-
-        # get the search range
-        rng = [value - tol, value + tol]
-
-        # wrap range if coordinate is cyclic
-        if c_info[0] == 'cyclic':
-            low = c_info[1][0]
-            upp = c_info[1][1]
-            if rng[0] < c_info[1][0] - atol:
-                rng[0] = (rng[0] - low) % (upp - low) + low
-            if rng[1] > c_info[1][1] + atol:
-                rng[1] = (rng[1] - low) % (upp - low) + low
-
-        # get the coordinates
-        coords = eval(f"self.{coordinate}")
-
-        # get the mask
-        if rng[0] <= rng[1]:
-            mask = (coords >= rng[0] - atol) & (coords <= rng[1] + atol)
-        else:
-            mask = (coords >= rng[0] - atol) | (coords <= rng[1] + atol)
-
-        # plot all and returned points
-        if show:
-            self.show(mask)
-
-        index = np.where(mask)
-
-        return index, mask
-
     def rotate(self, rotation: str, value=None, degrees=True, inverse=False):
         """
         Rotate points stored in the object around the origin of coordinates.
@@ -2529,53 +1451,6 @@ class Coordinates():
 
         return _systems
 
-    def _exist_system(self, domain=None, convention=None, unit=None):
-        """
-        Check if a coordinate system exists and throw an error if it does not.
-
-        The coordinate systems are defined in self._systems.
-
-        Parameters
-        ----------
-        domain : string
-            Specify the domain of the coordinate system, e.g., 'cart'.
-        convention : string
-            The convention of the coordinate system, e.g., 'top_colat'
-        units: string
-            The unit of the coordinate system (rad, deg, or met for radians,
-            degrees, or meters)
-        """
-
-        if domain is None:
-            raise ValueError('The domain must be specified')
-
-        # get available coordinate systems
-        systems = self._systems()
-
-        # check if domain exists
-        assert domain in systems or domain is None, \
-            f"{domain} does not exist. Domain must be one of the following: "\
-            f"{', '.join(list(systems))}."
-
-        # check if convention exists in domain
-        if convention is not None:
-            assert convention in systems[domain] or convention is None, \
-                f"{convention} does not exist in {domain}. Convention must "\
-                f"be one of the following: {', '.join(list(systems[domain]))}."
-
-        # check if units exist
-        if unit is not None:
-            # get first convention in domain
-            # (units are the same within a domain)
-            if convention is None:
-                convention = list(systems[domain])[0]
-
-            cur_units = [u[0][0:3] for u in
-                         systems[domain][convention]['units']]
-            assert unit in cur_units, \
-                f"{unit} does not exist in {domain} convention "\
-                f"Unit must be one of the following: {', '.join(cur_units)}."
-
     def _exist_coordinate(self, coordinate, unit):
         """
         Check if coordinate and unit exist.
@@ -2605,53 +1480,15 @@ class Coordinates():
              "self.systems() for a list of possible "
              "coordinates and units"))
 
-    def _make_system(self, domain=None, convention=None, unit=None):
-        """
-        Make and return class internal information about coordinate system.
-        """
-
-        # check if coordinate system exists
-        self._exist_system(domain, convention, unit)
-
-        # get the new system
-        system = self._systems()
-        if convention is None:
-            convention = list(system[domain])[0]
-        system = system[domain][convention]
-
-        # get the units
-        if unit is not None:
-            units = [units for units in system['units']
-                     if unit == units[0][0:3]]
-            units = units[0]
-        else:
-            units = system['units'][0]
-            unit = units[0][0:3]
-
-        # add class internal keys
-        system['domain'] = domain
-        system['convention'] = convention
-        system['unit'] = unit
-        system['units'] = units
-
-        return system
-
     def _set_points(self, x, y, z):
         """
-        Check points and convert to matrix.
+        Check input and store in private ``_x``, ``_y``, and ``_z`` properties.
 
         Parameters
         ----------
-        convert : boolean, optional
-            Set self._points if convert = True. Return points as
-            matrix otherwise. The default is False.
-        system: dict, optional
-            The coordinate system against which the range of the points are
-            checked as returned from self._make_system. If system = None
-            self._system is used.
-
-        Set self._points, which is an atleast_2d numpy array of shape
-        [L,M,...,N, 3].
+        x, y, z: number, array like
+            The x, y, and z coordinates. All input data must be broadcastable
+            to a common shape.
         """
         # cast to numpy array
         x = np.atleast_1d(np.asarray(x, dtype=np.float64))
